+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Snapshot tests with react-testing-library.
+
 ## [2.6.5] - 2019-02-14
 
 ## [2.6.4] - 2019-02-14
@@ -14,89 +18,123 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [2.6.3] - 2019-02-12
 
 ## [2.6.2] - 2019-02-09
+
 ### Changed
+
 - Add page padding in menu link.
 
 ## [2.6.1] - 2019-02-09
+
 ### Fixed
+
 - Remove `inheritComponent` from blocks.
 
 ## [2.6.0] - 2019-01-28
 
 ## [2.5.0] - 2019-01-18
+
 ### Changed
+
 - Update React builder to 3.x.
 
 ## [2.4.1] - 2019-01-14
+
 ### Changed
+
 - Remove `undefined` class.
-w
+
 ## [2.4.0] - 2019-01-09
+
 ### Changed
+
 - Bye `pages.json`! Welcome `store-builder`.
 
 ## [2.3.0] - 2018-12-20
+
 ### Added
+
 - Support to messages builder.
 
 ## [2.2.0] - 2018-12-18
+
 ### Changed
+
 - Add support to CSS modules.
 
 ## [2.1.0] - 2018-12-18
+
 ### Feature
+
 - Render nothing if there is no links to be displayed.
 
 ## [2.0.1] - 2018-11-30
+
 - Update link rendering to include internal/external.
 
 ## [2.0.0] - 2018-11-27
+
 ### Fixed
+
 - Use design tokens for typography and colors.
 
 ## [1.1.0] - 2018-10-31
+
 ### Changed
+
 - Move app to typescript.
 
 ## [1.0.4] - 2018-09-05
+
 ### Changed
+
 - Menu's schema now uses type array.
 
 ## [1.0.3] - 2018-08-30
+
 ### Fixed
+
 - Links default colors.
 
 ## [1.0.2] - 2018-08-24
+
 ### Changed
+
 - Change background color to white.
 
 ## [1.0.1] - 2018-08-24
+
 ### Changed
+
 - Change the layout to a light version.
 
 ## [1.0.0] - 2018-08-17
+
 ### Changed
+
 - remove border-bottom from `menu`.
 - decreased menu height.
 
 ## [0.3.0] - 2018-6-11
 
 ### Added
+
 - Support for internal and external link routes.
 - Internationalization to the schema.
 
 ## [0.2.0] - 2018-5-30
 
 ### Changed
+
 - Hide menu on mobile mode.
 
 ## [0.1.1] - 2018-05-07
 
-### Fixed 
-- Fix format text of left item. 
+### Fixed
+
+- Fix format text of left item.
 
 ## [0.1.0] - 2018-05-05
 
 ### Added
- - Initial files and logic created.
 
+- Initial files and logic created.